requirements.txt
boto3==1.35.99
botocore==1.35.99
click==8.1.8
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
rich==14.0.0
s3transfer==0.10.4
shellingham==1.5.4
six==1.17.0
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.13.2
urllib3==2.4.0
